// src/types/medical-unit.ts
enum MedicalUnit {
  capsule = 'capsule',
  tablet = 'tablet',
  injection = 'injection',
  spray = 'spray',
  drop = 'drop',
  suppository = 'suppository',
  sachet = 'sachet',
  unit = 'unit',
  vaginalCapsule = 'vaginalCapsule',
  mcg = 'mcg',
  mg = 'mg',
  g = 'g',
  ml = 'ml',
}

export default MedicalUnit;